backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
pymongo==4.6.1
Pillow==10.2.0
numpy==1.26.3
Werkzeug==3.0.1
torch>=2.1.0
torchvision>=0.16.0
sam2>=1.0
